+++ package.json
@@ -87,6 +87,6 @@
     "sass-loader": "^13.2.0",
     "storybook": "^7.0.0-rc.7",
     "tsx": "^3.12.3",
-    "typescript": "^4.9.5"
+    "typescript": "^5.0.2"
   }
 }
