+++ tsconfig.json
@@ -12,10 +12,9 @@
     "module": "esnext",
     "moduleResolution": "node",
     "resolveJsonModule": true,
-    "isolatedModules": true,
     "jsx": "preserve",
-    "importsNotUsedAsValues": "error",
-    "noUnusedParameters": true,
+    "verbatimModuleSyntax": true,
+    "noUnusedParameters": true
   },
   "include": ["next-env.d.ts", "src"],
   "exclude": ["node_modules", ".next"]
